feat: introduce write and admin permission levels for answer collections (#4529)

+++ packages/graphql/src/services/resources.ts
@@ -87,9 +87,11 @@ export async function createAnswerCollection(
     numSharedUsers: newCollection._count?.permissions,
     numOfEntries: newCollection._count.entries,
     isOwner: true,
+    isImported: false,
     isEditable: true,
-    isImported: newCollection.originalId !== null,
-    isAccessGranted: false,
+    isShareable: true,
+    isRemovable: true,
+    isDeletionAllowed: true,
   }
 }
 
@@ -278,16 +280,18 @@ export async function getSingleAnswerCollection(
       isEditable: true,
     }
   } else {
+    const accessLevel = collection.permissions[0]?.accessLevel
     return {
       ...collection,
       accessType: AccessType.SHARED,
       sharingStatus: DB.PermissionStatus.GRANTED, // need to be granted for query above to succeed
-      sharingLevel:
-        collection.permissions[0]?.accessLevel ?? DB.AccessLevel.READ,
+      sharingLevel: accessLevel ?? DB.AccessLevel.READ,
       ownerShortname: collection.owner?.shortname,
       catalogCollectionId,
       isOwner: false,
-      isEditable: false,
+      isEditable:
+        accessLevel === DB.AccessLevel.WRITE ||
+        accessLevel === DB.AccessLevel.ADMIN,
     }
   }
 }
@@ -299,20 +303,19 @@ export async function getAnswerCollectionPermissions(
   const collection = await ctx.prisma.answerCollection.findUnique({
     where: {
       id: collectionId,
-      // TODO: only allow access to owners or users with granted admin access
       OR: [
         {
           ownerId: ctx.user.sub,
         },
-        // {
-        //   permissions: {
-        //     some: {
-        //       userId: ctx.user.sub,
-        //       permissionStatus: DB.PermissionStatus.GRANTED,
-        //       accessLevel: DB.AccessLevel.ADMIN,
-        //     },
-        //   },
-        // },
+        {
+          permissions: {
+            some: {
+              userId: ctx.user.sub,
+              permissionStatus: DB.PermissionStatus.GRANTED,
+              accessLevel: DB.AccessLevel.ADMIN,
+            },
+          },
+        },
       ],
     },
     include: {
@@ -370,29 +373,16 @@ export async function shareAnswerCollection(
   },
   ctx: ContextWithUser
 ) {
-  // fetch collection and verify that the user has permissions to share it
-  const collection = await ctx.prisma.answerCollection.findUnique({
-    where: {
-      id: collectionId,
-      OR: [
-        {
-          ownerId: ctx.user.sub,
-        },
-        // TODO: update with admin permissions here as soon as they are implemented
-        // {
-        //   permissions: {
-        //     some: {
-        //       userId: ctx.user.sub,
-        //       permissionStatus: DB.PermissionStatus.GRANTED,
-        //       accessLevel: DB.AccessLevel.ADMIN,
-        //     },
-        //   },
-        // },
-      ],
+  // verify that user has either owner or admin access
+  const valid = await validateCollectionPermissions(
+    {
+      collectionId,
+      acceptedAccessLevels: [DB.AccessLevel.ADMIN],
     },
-  })
+    ctx
+  )
 
-  if (!collection) {
+  if (!valid) {
     return null
   }
 
@@ -477,6 +467,72 @@ export async function shareAnswerCollection(
   }
 }
 
+export async function changeCollectionAccessLevel(
+  {
+    collectionId,
+    permissionId,
+    accessLevel,
+  }: {
+    collectionId: number
+    permissionId: number
+    accessLevel: DB.AccessLevel
+  },
+  ctx: ContextWithUser
+) {
+  // verify that user has either owner or admin access
+  const valid = await validateCollectionPermissions(
+    {
+      collectionId,
+      acceptedAccessLevels: [DB.AccessLevel.ADMIN],
+    },
+    ctx
+  )
+
+  if (!valid) {
+    return null
+  }
+
+  // update the access level of the permission
+  const permission = await ctx.prisma.permission.update({
+    where: {
+      id: permissionId,
+    },
+    data: {
+      accessLevel,
+    },
+    include: {
+      user: {
+        select: {
+          id: true,
+          shortname: true,
+          email: true,
+        },
+      },
+    },
+  })
+
+  // if the permission did not exist in the first place, return null
+  if (!permission) {
+    return null
+  }
+
+  // invalidate permission
+  ctx.emitter.emit('invalidate', {
+    typename: 'Permission',
+    id: permission.id,
+  })
+
+  return {
+    permissionId: permission.id,
+    userId: permission.user?.id,
+    username: permission.user?.shortname,
+    userEmail: permission.user?.email,
+    userGroupId: undefined,
+    userGroupName: undefined,
+    accessLevel: permission.accessLevel,
+  }
+}
+
 export async function getAnswerCollectionsInfo(ctx: ContextWithUser) {
   const user = await ctx.prisma.user.findUnique({
     where: {
@@ -547,11 +603,11 @@ export async function getAnswerCollectionsInfo(ctx: ContextWithUser) {
     numSharedUsers: collection._count?.permissions,
     numOfEntries: collection._count.entries,
     isOwner: true,
-    isEditable: true,
-    isRemovable: collection._count?.linkedElements === 0,
     isImported: collection.originalId !== null,
-    isShareable: true, // owned answer collections can always be shared
-    isAccessGranted: false,
+    isEditable: true, // owner always has editing permissions
+    isShareable: true, // owner always has sharing permissions
+    isRemovable: collection._count.linkedElements === 0, // can be removed from own account if not linked to any elements
+    isDeletionAllowed: true, // owner always has deletion objects
   }))
 
   const sharedCollections = user.sharedObjects.flatMap((object) => {
@@ -569,11 +625,13 @@ export async function getAnswerCollectionsInfo(ctx: ContextWithUser) {
       sharingLevel: object.accessLevel,
       ownerShortname: collection.owner?.shortname,
       isOwner: false,
-      isEditable: false,
-      isRemovable: collection._count.linkedElements === 0,
-      isImported: false,
+      isImported: false, // shared objects cannot be imported
+      isEditable:
+        object.accessLevel === DB.AccessLevel.WRITE ||
+        object.accessLevel === DB.AccessLevel.ADMIN, // only users with write or admin access can edit shared answer collections
       isShareable: object.accessLevel === DB.AccessLevel.ADMIN, // only users with admin access can share a shared answer collection
-      isAccessGranted: true, // requested ansewr collections are not returned by this query
+      isRemovable: collection._count.linkedElements === 0, // can be removed from own account if not linked to any elements
+      isDeletionAllowed: object.accessLevel === DB.AccessLevel.ADMIN, // only users with admin access can delete answer collections
     }
   })
 
@@ -878,16 +936,70 @@ async function incrementCollectionVersion(
   return collection
 }
 
+async function validateCollectionPermissions(
+  {
+    collectionId,
+    acceptedAccessLevels,
+  }: {
+    collectionId: number
+    acceptedAccessLevels: DB.AccessLevel[]
+  },
+  ctx: ContextWithUser
+) {
+  const collection = await ctx.prisma.answerCollection.findUnique({
+    where: {
+      id: collectionId,
+      OR: [
+        {
+          ownerId: ctx.user.sub,
+        },
+        {
+          permissions: {
+            some: {
+              userId: ctx.user.sub,
+              permissionStatus: DB.PermissionStatus.GRANTED,
+              accessLevel: {
+                in: acceptedAccessLevels,
+              },
+            },
+          },
+        },
+      ],
+    },
+  })
+
+  if (!collection) {
+    return false
+  }
+
+  return true
+}
+
 export async function editAnswerCollectionEntry(
   {
     id,
     value,
+    collectionId,
   }: {
     id: number
     value: string
+    collectionId: number
   },
   ctx: ContextWithUser
 ) {
+  // verify that the user has at least writer permissions for the collection
+  const valid = await validateCollectionPermissions(
+    {
+      collectionId,
+      acceptedAccessLevels: [DB.AccessLevel.WRITE, DB.AccessLevel.ADMIN],
+    },
+    ctx
+  )
+
+  if (!valid) {
+    return null
+  }
+
   // update entry in the database
   const updatedEntry = await ctx.prisma.answerCollectionEntry.update({
     where: {
@@ -908,9 +1020,22 @@ export async function editAnswerCollectionEntry(
 }
 
 export async function deleteAnswerCollectionEntry(
-  { id }: { id: number },
+  { id, collectionId }: { id: number; collectionId: number },
   ctx: ContextWithUser
 ) {
+  // verify that the user has at least writer permissions for the collection
+  const valid = await validateCollectionPermissions(
+    {
+      collectionId,
+      acceptedAccessLevels: [DB.AccessLevel.WRITE, DB.AccessLevel.ADMIN],
+    },
+    ctx
+  )
+
+  if (!valid) {
+    return null
+  }
+
   // delete answer option from the database
   const updatedEntry = await ctx.prisma.answerCollectionEntry.delete({
     where: {
@@ -937,6 +1062,19 @@ export async function addAnswerCollectionOption(
   },
   ctx: ContextWithUser
 ) {
+  // verify that the user has at least writer permissions for the collection
+  const valid = await validateCollectionPermissions(
+    {
+      collectionId,
+      acceptedAccessLevels: [DB.AccessLevel.WRITE, DB.AccessLevel.ADMIN],
+    },
+    ctx
+  )
+
+  if (!valid) {
+    return null
+  }
+
   // add new answer option to the database
   const newEntry = await ctx.prisma.answerCollectionEntry.create({
     data: {
@@ -1420,7 +1558,7 @@ export async function resolveObjectSharingRequest(
     where: {
       id: permissionId,
       userId,
-      accessLevel: accessLevel ?? DB.AccessLevel.READ,
+      accessLevel: DB.AccessLevel.READ, // access requests are always assigned read access level
       permissionStatus: DB.PermissionStatus.REQUESTED,
       objectOwnerId: ctx.user.sub,
     },
@@ -1438,6 +1576,7 @@ export async function resolveObjectSharingRequest(
       },
       data: {
         permissionStatus: DB.PermissionStatus.GRANTED,
+        accessLevel,
       },
     })
   } else {
